Added option to enable/disable hardlinking, option to avoid hardlinking of specific file endings

+++ daemon/nntp/ArticleWriter.cpp
@@ -583,7 +583,7 @@ void ArticleWriter::CompleteFileParts()
 		GuardedDownloadQueue guard = DownloadQueue::Guard();
 		m_fileInfo->SetCrc(crc);
 
-		if (m_fileInfo->IsHardLinked())
+		if (m_fileInfo->IsHardLinked() && Util::EmptyStr(g_Options->GetInterDir()))
 		{
 			// No need to rename, remove hardlink of .out.tmp-file
 			if (FileSystem::DeleteFile(m_outputFilename.c_str()))
